+++ lib/dexter/client.rb
@@ -29,6 +29,7 @@ module Dexter
     dexter [options]
 
 Options:)
+        o.boolean "--analyze", "analyze tables that haven't been analyzed in the past hour", default: false
         o.boolean "--create", "create indexes", default: false
         o.array "--exclude", "prevent specific tables from being indexed"
         o.string "--include", "only include specific tables"